feat: Update color and style in Start screen
The code changes include updating the color and style in the Start screen. The color of the input card is changed to a lighter gray (#bfbdbd) and the text color of the header is changed to blue (#0000ff). Additionally, a button container is added to the screen to display the submit button and another button.